ksp0607: fix fec_imx probe deferred permanently
The reason the probe of the network is deferred permanently was that
the phy-supply regulator is requested, but has no driver for the PMIC,
but a driver as the MFD device. This will return -EPROBE_DEFER.
(The da9063/2 does not implement the regulator API device of MFD).

The imx6dl-phy-core device tree fixes the issue by deleting the phy-supply dt
node.

The solution chosen here is to use the special dt property, to mark
a regulator (or group) to use the 'dummy-regulator' (aka none) to
fullfill the requirement. This way the probe can succeed.

The accesss to the regulator itself is not really needed because it
already should be in the correct state (enabled).